requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
langchain==0.1.0
pydantic==2.6.0
python-dotenv==1.0.0
psutil
astor